move agg to root module
