+++ src/expr/src/agg/string_agg.rs
@@ -23,7 +23,7 @@ use risingwave_common::util::iter_util::ZipEqFast;
 use risingwave_common::util::memcmp_encoding;
 use risingwave_common::util::sort_util::{ColumnOrder, OrderType};
 
-use crate::vector_op::agg::aggregator::Aggregator;
+use super::Aggregator;
 use crate::{ExprError, Result};
 
 #[derive(Clone)]
