--- src/expr/src/vector_op/agg/mod.rs
@@ -1,27 +1,0 @@
-// Copyright 2023 RisingWave Labs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-mod aggregator;
-mod approx_count_distinct;
-mod array_agg;
-mod count_star;
-mod filter;
-mod functions;
-mod general_agg;
-mod general_distinct_agg;
-mod general_sorted_grouper;
-mod string_agg;
-
-pub use aggregator::{create_agg_state_unary, AggStateFactory, BoxedAggState};
-pub use general_sorted_grouper::{create_sorted_grouper, BoxedSortedGrouper, EqGroups};
